Combine export button components and use Button Component
