+++ src/pages/dashboard/inquiry-builder/tabs/analysis/index.tsx
@@ -9,8 +9,7 @@ import React from 'react';
 import PerQuestionTab from './per-question';
 import PerResponseTab from './per-response';
 import ViaChatTab from './via-chat';
-import ExportCSV from './export-csv';
-import ExportJSON from './export-json';
+import ExportButton from './export-button';
 
 interface AnalysisTabProps {
   id: string;
@@ -49,10 +48,12 @@ const AnalysisTab: React.FC<AnalysisTabProps> = ({ id }) => {
 
   return (
     <div className="mt-8 rounded-2xl">
-      <div className="flex justify-end items-center gap-4 mb-4">
-        <ExportCSV id={id} responses={inquiryResponseData?.getInquiryResponses ?? []} />
-        <ExportJSON id={id} responses={inquiryResponseData?.getInquiryResponses ?? []} />
-      </div>
+      {(inquiryResponseData?.getInquiryResponses?.length ?? 0) > 0 && (
+        <div className="flex justify-end items-center gap-4 mb-4">
+          <ExportButton id={id} responses={inquiryResponseData?.getInquiryResponses ?? []} type="csv" />
+          <ExportButton id={id} responses={inquiryResponseData?.getInquiryResponses ?? []} type="json" />
+        </div>
+      )}
 
       <TabGroup>
         <TabList className="flex space-x-1 rounded-xl border-2 border-white mb-4">
